create Bid component to separate bid logic into one component
